Move client initialization to separate file
